% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_auxiliaryFunctions.R
\name{acomb}
\alias{acomb}
\title{Combine matrices along in a three dimensional matrix}
\usage{
acomb(...)
}
\arguments{
\item{...}{arrays of dimension 3}
}
\description{
Combine matrices along in a three dimensional matrix
}
