% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_Argument.R
\name{ArgumentSS}
\alias{ArgumentSS}
\title{Class definition
It should be used when vectors or matrices are to be substacked}
\usage{
ArgumentSS(dims, x)
}
\arguments{
\item{dims}{Dimensions of the argument}

\item{x}{Vector of names that can be used to be assigned to the columns/rows}
}
\description{
Class definition
It should be used when vectors or matrices are to be substacked
}
