% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_Network.R
\name{BlockGraph}
\alias{BlockGraph}
\title{Creates object of class \code{BlockGraph}}
\usage{
BlockGraph(x, data)
}
\arguments{
\item{x}{the graph, an object of class \code{igraph} with named nodes}

\item{data}{dataset with named columns}
}
\description{
Creates object of class \code{BlockGraph}
}
\examples{
g<- graph(c(1,2,2,3), directed = FALSE)
g<- set.vertex.attribute(g, "name", V(g), c("a", "b", "c")) # name the nodes of the graph
data<- matrix(rnorm(10*3), 10, 3)
colnames(data)<- c("a", "b", "c") # name the columns of the data
bg<- BlockGraph(x = g, data = data)
}
