% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_confInt.R
\name{confInt}
\alias{confInt}
\alias{confInt.EKS}
\title{Confidence intervals for pairwise EC estimates}
\usage{
confInt(obj, ...)

\method{confInt}{EKS}(obj, evpo, k, level = 0.05, ...)
}
\arguments{
\item{obj}{is an object of class EKS which contains the estimates of the edge weights}

\item{...}{additional arguments}

\item{evpo}{is the matrix of evaluation points used in the EC estimator. It should contain only pairs of ones.}

\item{k}{is the number of upper order statistics. It should be the same as in the estimates of the edge weights.}

\item{level}{is the level of confidence. default is 0.05.}
}
\description{
Provides confidence intervals for the estimates using the ECE. It is suitable when only pairs are used -
pairwise extremal coefficients.
}
\examples{
# See Vignette "Application Seine"
}
