% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_coordinates.R
\name{Coordinates}
\alias{Coordinates}
\title{Creates an empty object of class \code{Coordinates}}
\usage{
Coordinates()
}
\description{
It is used for creating the coordinates where the stable tail dependence function (stdf) is evaluated.
For instance in the extremal coefficient estimator. For more information about it see
Vignettes "Estimation - Note 3" and "Code - Note 3" and "Subsets and Coordinates".
}
