% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_Network.R
\name{CovSelectTree}
\alias{CovSelectTree}
\title{Creates object of class \code{CovSelectTree}}
\usage{
CovSelectTree(g, Data)
}
\arguments{
\item{g}{should be the graph, an object of class \code{igraph} with named nodes}

\item{Data}{is the dataset with named columns}
}
\description{
Creates object of class \code{CovSelectTree}
}
\examples{
g<- graph(c(1,2,2,3), directed=FALSE)
g<- set.vertex.attribute(g, "name", V(g), c("a", "b", "c")) # name the nodes of the graph
data<- matrix(rnorm(10*3), 10,3)
colnames(data)<- c("a", "b", "c") # name the columns of the data
cst<- CovSelectTree(g = g, Data = data)
}
