% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_dataSets.R
\docType{data}
\name{Danube}
\alias{Danube}
\title{Danube Dataset}
\description{
Dataset of events on Upper Danube. For detailed description of the data please read in the Appendix
of the paper below.
}
\references{
Peiman Asadi. Anthony C. Davison. Sebastian Engelke. "Extremes on river networks."
Ann. Appl. Stat. 9 (4) 2023 - 2050, December 2015. https://doi.org/10.1214/15-AOAS863
}
\author{
Sebastian Engelke and Adrian Hitz
}
\keyword{data}
