% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_dataSets.R
\docType{data}
\name{DanubeTree}
\alias{DanubeTree}
\title{Danube tree}
\value{
The graph used in the reference below. An \code{igraph} object.
}
\description{
Tree corresponding to the Danube network. It consists of 31 nodes.
}
\references{
Peiman Asadi. Anthony C. Davison. Sebastian Engelke. "Extremes on river networks."
Ann. Appl. Stat. 9 (4) 2023 - 2050, December 2015. https://doi.org/10.1214/15-AOAS863
}
\author{
Peiman Asadi, Anthony C. Davison, Sebastian Engelke.
}
\keyword{data}
