% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_diagnost.R
\name{diagnost}
\alias{diagnost}
\alias{diagnost.HRMnetwork}
\title{Performs diagnosis on the generated data of object \code{HRMtree}}
\usage{
diagnost(obj, ...)

\method{diagnost}{HRMnetwork}(obj, X, id, ...)
}
\arguments{
\item{obj}{Object of class \code{HRMtree}}

\item{X}{A matrix of data to be diagnosted}

\item{id}{The name of the variable for which the diagnostics is done.}
}
\description{
As one diagnostic tool, the function produces a graph comparing the real marginal distribution
with the empirical distribution of the variable in \code{id}. As a second diagnostic tool
for every adjacent node to \code{id} it computes the real and the empirical copulas at coordinates
(0.5,0.5).
}
\examples{
# create a graph with named vertices
g<- graph(c("a", "b", "b","c", "b", "d"), directed = FALSE)
# create a HRMtree object
myobj<- HRMtree(g)
x<- c(0.1,0.2,0.3)
myobj<- setParams(myobj, x)
# create a dataset
mydata<- rHRMtree(myobj, 1000)
# do diagnostic on the node "b"
diagnost(myobj, mydata, "b")
# include noise in the data
mydata<- rHRMtree(myobj, 1000, noise = TRUE)
diagnost(myobj, mydata, "c")
}
