% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_edgeNameAlongPath.R
\name{edge_names_along_path}
\alias{edge_names_along_path}
\alias{edge_names_along_path.HRMnetwork}
\title{Obtains the edge names along a path between two vertices on a tree}
\usage{
edge_names_along_path(obj, ...)

\method{edge_names_along_path}{HRMnetwork}(obj, rt, id, edge_names = TRUE, ...)
}
\arguments{
\item{obj}{object of class \code{HRMnetwork}}

\item{...}{additional arguments}

\item{rt}{the name of the node from which the path is requested}

\item{id}{The name of the node to which the path is requested}

\item{edge_names}{Returns a vector with the names of the edges. Default is TRUE. If FALSE, it returns
a named vector of ones and zeros with ones if the corresponding edge is part of the path}
}
\value{
The names of the edges along the path between \code{rt} and \code{id}
}
\description{
Obtains the edge names along a path between two vertices on a tree
}
