% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{EKS}
\alias{EKS}
\title{\code{EKS} class}
\usage{
EKS(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{EKS} with two slots: one containing the graph and the other the edge weights
corresponding to the dependence parameters.
}
\description{
\code{EKS} class
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
eks<- EKS(g)
}
