% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{EKS_part}
\alias{EKS_part}
\title{\code{EKS_part} class}
\usage{
EKS_part(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{EKS_part} with two slots: one containing the graph and the other the edge weights
corresponding to the dependence parameters.
}
\description{
\code{EKS_part} class
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
ekspart<- EKS_part(g)
}
