% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{EngHitz}
\alias{EngHitz}
\title{\code{EngHitz} class}
\usage{
EngHitz(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{Enghitz} with two slots: one containing the graph and the other the edge weights
corresponding to the dependence parameters.
}
\description{
\code{EngHitz} class
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
enghitz<- EngHitz(g)
}
