% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_estimate.R
\name{estimate.EKS_part}
\alias{estimate.EKS_part}
\title{EKS estimator based on subsets}
\usage{
\method{estimate}{EKS_part}(obj, Data, subsets, k_ratio, obj2, Ubar, par_names, gr, xx, ...)
}
\description{
EKS estimator based on subsets
}
