% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_estimate.R
\name{estimate.MLE1}
\alias{estimate.MLE1}
\title{Estimates the parameters of a HRM tree using the MLE version 1}
\usage{
\method{estimate}{MLE1}(obj, Data, subsets, k_ratio, obj2, Ubar, par_names, gr, ...)
}
\arguments{
\item{obj2}{Object of class \code{Gtree}}

\item{Ubar}{The set of nodes for which there are no data available}

\item{par_names}{A character vector with the names of the parameters associated to the edges of the tree}

\item{gr}{The graph associated to the mle1obj}

\item{mle1obj}{Object of class \code{MLE1}}
}
\description{
The function is designed to be used as inherited method
}
\references{
Asenova, S and Segers, J. Husler-Reiss Markov tree
}
