% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_flowConnection.R
\name{flowConnection}
\alias{flowConnection}
\alias{flowConnection.FlowConnectionMatrix}
\alias{flowConnection.FlowConnectionGraph}
\title{Generic for \code{flowConnection}}
\usage{
flowConnection(obj, ...)

\method{flowConnection}{FlowConnectionMatrix}(obj, g, vertices = rownames(obj), ...)

\method{flowConnection}{FlowConnectionGraph}(obj, vertices = get.vertex.attribute(obj, "name", V(obj)), ...)
}
\description{
Generic for \code{flowConnection}
}
