% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_flowConnectionInput.R
\name{FlowConnectionGraph}
\alias{FlowConnectionGraph}
\title{Object of class \code{FlowConnectionGraph}}
\usage{
FlowConnectionGraph(g)
}
\arguments{
\item{g}{should be an object of class "igraph"}
}
\description{
It is used as an input to create subsets based on the criterion of flow connectedness.
Within a subset all nodes will be flow connected.
}
\examples{
seg_dir<- graph(c("2","paris", "meaux", "2", "melun", "2","5", "melun","nemours", "5", "sens", "5"),
 directed = TRUE)
fcg<- FlowConnectionGraph(seg_dir)
}
