% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_fnk.R
\name{fnk}
\alias{fnk}
\title{A function combining the empirical and the parametric stdf}
\usage{
fnk(params, eksobj, evalPoints, stdf_emp, Ubar)
}
\arguments{
\item{params}{A vector of |V|-1 parameters' values}

\item{eksobj}{An object of class \code{EKS}}

\item{evalPoints}{A matrix of evaluation points with column names according to the edge names}

\item{stdf_emp}{The empirical stdf, evaluated at the 'evalPoints', a vector of values}
}
\description{
Combines the empirical and the parametric stdf of a HRM tree in
\eqn{||l_{kn}-l(\theta)||^2}. It is the function which is minimised to obtain the EKS estimators
of the dependence parameters \eqn{\theta}.
}
