% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_for_u_in_U.R
\name{for_u_in_U}
\alias{for_u_in_U}
\alias{for_u_in_U.Argument}
\title{Generic for method \code{for_u_in_U}}
\usage{
for_u_in_U(obj, ...)

\method{for_u_in_U}{Argument}(obj, obj2, subsets, k_ratio, h1, depParams, Ubar = NULL, ...)
}
\arguments{
\item{obj}{Object of class \code{Argument}}

\item{obj2}{Object of class \code{HRMtree, GTree, CovSelectTree}}

\item{subsets}{Object of class \code{RootDepSet}}

\item{k_ratio}{The fraction of the upper order statistics in the sample size}

\item{h1}{A numeric vector with the length of the subsets for each root, after removing the corresponding root and the nodes with missing data.}

\item{depParams}{Named vector with the values with the values of the parameters}

\item{Ubar}{Vector with the names of the nodes with missing data}
}
\description{
It generates arguments that are used in the estimation via MM and ML methods. For every node \eqn{u\in U} produces the empirical or the parameric covariance matrices and
depending on the class of the first argument it combines them in (so far) three different ways:
- substacking of vectors
- substacking of matrices
- creating a block diagonal matrix
}
