% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_getters.R
\name{getData}
\alias{getData}
\alias{getData.default}
\alias{getData.Network}
\title{It retrieves the dataset from an appropriate object}
\usage{
getData(obj, ...)

\method{getData}{default}(obj, ...)

\method{getData}{Network}(obj, ...)
}
\arguments{
\item{obj}{Object with slot 'data'}

\item{...}{Additional arguments}
}
\description{
The function is applied on objects of appropriate class with slot 'data'.
It is used instead of the command 'obj_name$data'.
}
