% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_getters.R
\name{getGraph}
\alias{getGraph}
\alias{getGraph.default}
\alias{getGraph.Network}
\alias{getGraph.HRMnetwork}
\title{It retrives the graph of an object of appropriate class}
\usage{
getGraph(obj, ...)

\method{getGraph}{default}(obj, ...)

\method{getGraph}{Network}(obj, ...)

\method{getGraph}{HRMnetwork}(obj, ...)
}
\arguments{
\item{obj}{Object of appropriate class ('Network', 'HRMnetwork')}
}
\description{
The function is applied on objects of appropriate class with slot 'graph'. It is used instead of the command 'obj_name$graph'.
}
