% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_getters.R
\name{getNoDataNodes}
\alias{getNoDataNodes}
\alias{getNoDataNodes.default}
\alias{getNoDataNodes.Network}
\title{It retrives the nodes without data}
\usage{
getNoDataNodes(obj, ...)

\method{getNoDataNodes}{default}(obj, ...)

\method{getNoDataNodes}{Network}(obj, ...)
}
\arguments{
\item{obj}{Object of class \code{Network} or its subclasses \code{Tree, GTree, CovSelectTree, BlockGraph}}

\item{...}{additional arguments}
}
\description{
It retrieves the value of the slot \code{$noDataNodes}, the set of nodes for which there are no data available
 for an object of class \code{Network}.
}
