% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_getters.R
\name{getNodesWithData}
\alias{getNodesWithData}
\alias{getNodesWithData.default}
\alias{getNodesWithData.Network}
\title{It retrieves the nodes for which data are missing}
\usage{
getNodesWithData(obj, ...)

\method{getNodesWithData}{default}(obj, ...)

\method{getNodesWithData}{Network}(obj, ...)
}
\arguments{
\item{obj}{Object of class 'Network'}
}
\description{
It retrieves the slot 'nodesWithData' of an object 'Network'. The slot represents the set of nodes for
which there is no data available.
}
