% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_getters.R
\name{getParams}
\alias{getParams}
\alias{getParams.default}
\alias{getParams.HRMnetwork}
\title{It retrives the parameters associated to an HRMnetwork}
\usage{
getParams(obj, ...)

\method{getParams}{default}(obj, ...)

\method{getParams}{HRMnetwork}(obj, ...)
}
\arguments{
\item{obj}{An object of class 'HRMnetwork'}
}
\description{
It retrieves the value of the slot 'depParams' of a HRMnetwork.
}
