% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_getters.R
\name{getRoot}
\alias{getRoot}
\alias{getRoot.default}
\alias{getRoot.RootDepSet}
\title{It retrieves the roots associated to a subset}
\usage{
getRoot(obj, ...)

\method{getRoot}{default}(obj, ...)

\method{getRoot}{RootDepSet}(obj, ...)
}
\arguments{
\item{obj}{An object of class 'RootDepSet'}
}
\description{
It retrieves the slot 'root'
}
