% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_getters.R
\name{getValue}
\alias{getValue}
\alias{getValue.default}
\alias{getValue.RootDepSet}
\title{It retrieves the subsets of an object 'RootDepSet'}
\usage{
getValue(obj, ...)

\method{getValue}{default}(obj, ...)

\method{getValue}{RootDepSet}(obj, ...)
}
\arguments{
\item{obj}{An object of class 'RootDepSet'}

\item{...}{additional arguments}
}
\description{
It retrieves the slot 'value' of an object of class 'RootDepSet'.
}
