% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_gremes.R
\docType{package}
\name{gremes-package}
\alias{gremes}
\alias{gremes-package}
\title{gremes: Package for estimation of extremal dependence on graphical models}
\description{
The package provides three estimators of the dependence parameters of a
}
\details{
These are details
}
\section{Most important functions }{

kjmuhmoihùo
}

\section{La-la a }{


lkuliumiuhmikn,
}

\references{
Asenova, S., Mazo, G., Segers, J. Extremal dependence in Hüsler-Reiss Markov tree.

Segers, J. Regularly varying Markov trees
}
\author{
Stefka Asenova
}
