% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_h.R
\name{h}
\alias{h}
\alias{h.RootDepSet}
\title{Number of PoT variables in a subset}
\usage{
h(obj, ...)

\method{h}{RootDepSet}(obj, U_bar = NULL)
}
\arguments{
\item{obj}{Object of class \code{RootDepSet}}

\item{U_bar}{The set of nodes for which data are missing, as character vector}
}
\description{
Computes the h_u, which is
the number of the elements in the subset after substracting the root and the nodes for which data
are missing
}
