% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{HRMBG}
\alias{HRMBG}
\title{\code{HRMBG} class}
\usage{
HRMBG(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{HRMBG} with two slots: one containing the graph and the other the edge weights
corresponding to the dependence parameters.
}
\description{
\code{HRMBG} class
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
hrmbg<- HRMBG(g)
}
