% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{HRMnetwork}
\alias{HRMnetwork}
\title{\code{HRMnetwork} class}
\usage{
HRMnetwork(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{HRMnetwork} with two slots: one containing the graph and the other the edge weights
corresponding to the dependence parameters.
}
\description{
\code{HRMnetwork} class
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
hrmnet<- HRMnetwork(g)
}
