% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_intersect.R
\name{intersect.list}
\alias{intersect.list}
\title{Intersects the elements in a list; the list can have more than 2 elements}
\usage{
\method{intersect}{list}(x)
}
\description{
Intersects the elements in a list; the list can have more than 2 elements
}
