% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_ldParams.R
\name{ldParams.HRMtree}
\alias{ldParams.HRMtree}
\title{Vector of derivatives of the stdf
Computes the vector of partial derivatives of the stdf with respect to the vector of parameters theta in the HRM tree.}
\usage{
\method{ldParams}{HRMtree}(obj, x, ...)
}
\arguments{
\item{obj}{Object of a suitable class}

\item{x}{vector of coordinates, the function is made when x contains only two positive coordinates.}
}
\value{
A vector of length |E|. Each element of it represents the derivative of \eqn{l(e_J; \theta)} with respect
to the parameter \eqn{\theta} associated to a particular edge.
}
\description{
Vector of derivatives of the stdf
Computes the vector of partial derivatives of the stdf with respect to the vector of parameters theta in the HRM tree.
}
