% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_ll.R
\name{llmle2}
\alias{llmle2}
\title{Log-likelihood function for estimation of the second type of MLE}
\usage{
llmle2(params, data, obj, subsets, k_ratio, h1, Ubar)
}
\description{
Produces the log-likelihood fuction for estimation of the second type MLE
 @param params The parameters to estimate, It should be a named vector of initial values
 named according to the names of the edges in the
 whole graph (!!! Named vector of values)
 @param data The dataset. It should be a matrix with dimensions k by sum(h).
 @param obj Object of class HRMtree. This one is necessary for construction
 of the covariance matrix, see details
 @details The covariance matrix constructed in this function and passed to the
 multivariate normal density is a block diagonal matrix. Each matrix on the diagonal
 corresponds to the covariance matrix of the variables \eqn{R_{W_u}}. The mean vector is equal
 to \emph{-diag(sig)/2}, where \emph{sig}  is the covariance matrix.
 @param subsets Object of class \code{RootDepSet}
 @param k_ratio The share of the upper order statistics
 @param h1 The number of the variables \eqn{R_{W_u}} for every set in subsets, after subtracting
 the root and the variables with missing data.
 @param Ubar the set of missing data nodes
 @importFrom mvtnorm dmvnorm
}
