% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_auxiliaryFunctions.R
\name{means_vars}
\alias{means_vars}
\title{Computes means and standard deviations for a matrix of estimates}
\usage{
means_vars(obj)
}
\arguments{
\item{obj}{A matrix of dimensions (#parameters)X(#k)X(#estimators)X(#simulations)}
}
\value{
A list with two sub-lists: one containing a matrix of means (m), the other containing
a matrix of variances (v). The matrix of means is of dimensions (#estimators)X(#k)X(#parameters).
To access the estimates of the first parameter use [,,1] and so on. Similarly for the matrix
of variances.
}
\description{
It cleans up the matrix of estimates from NaNs and after that it computes the means
and the variances of the estimators per parameter and per k.
}
