% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{MLE}
\alias{MLE}
\title{\code{MLE} class}
\usage{
MLE(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{MLE} with two slots: one containing the graph and the other the edge weights
corresponding to the dependence parameters.
}
\description{
\code{MLE} class
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
mle<- MLE(g)
}
