% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{MLE1}
\alias{MLE1}
\title{\code{MLE1} class}
\usage{
MLE1(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{MLE1} with two slots: one containing the graph and the other the edge weights
corresponding to the dependence parameters.
}
\description{
\code{MLE1} class
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
mle1<- MLE1(g)
}
