% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{MLE2}
\alias{MLE2}
\title{\code{MLE2} class}
\usage{
MLE2(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{MLE2} with two slots: one containing the graph and the other the edge weights
corresponding to the dependence parameters.
}
\description{
\code{MLE2} class
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
mle2<- MLE2(g)
}
