% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{MLEave}
\alias{MLEave}
\title{\code{MLEave} class}
\usage{
MLEave(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{MLEave} with two slots: one containing the graph and the other the edge weights
corresponding to the dependence parameters.
}
\description{
\code{MLEave} class
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
mleave<- MLEave(g)
}
