% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{MME}
\alias{MME}
\title{\code{MME} class}
\usage{
MME(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{MME} with two slots: one containing the graph and the other the edge weights
corresponding to the dependence parameters.
}
\description{
\code{MME} class
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
mme<- MME(g)
}
