% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{MMEave}
\alias{MMEave}
\title{\code{MMEave} class}
\usage{
MMEave(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{MMEave} with two slots: one containing the graph and the other the edge weights
corresponding to the dependence parameters.
}
\description{
\code{MMEave} class
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
mmeave<- MMEave(g)
}
