% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_Network.R
\name{Network}
\alias{Network}
\title{Creates objects of calss \code{Network}}
\usage{
Network(x, data)
}
\arguments{
\item{x}{The graph with named nodes, an object of class \code{igraph}}

\item{data}{Data set, with named columns}
}
\description{
Creates objects of calss \code{Network}
}
\examples{
g<- graph(c(1,2,2,3), directed=FALSE)
g<- set.vertex.attribute(g, "name", V(g), c("a", "b" ,"c")) # name the nodes
data<- matrix(rnorm(10*3), 10,3)
colnames(data)<- c("a", "b", "c")  # name the columns
net<- Network(x = g, data = data)
}
