% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_plot.R
\name{plot.Simulation}
\alias{plot.Simulation}
\title{Plot method for Simulation results.}
\usage{
\method{plot}{Simulation}(x, ind_params, path_to_file = NULL, name_file = "plot", ...)
}
\arguments{
\item{ind_params}{a numeric vector of indices which indicates for which parameters the plots
are to be produced. For instance: the object that contains the estimates has four dimensions,
the first one equal to the number of parameters in the model. If ind_params=c(1,3) then plots
will be produced for the parameters in the first and third places of the first dimension
of the object with estimates.}

\item{obj}{an object of class Simulation}
}
\description{
Plot method for Simulation results.
}
