% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_auxiliaryFunctions.R
\name{plotEC}
\alias{plotEC}
\title{Plots Extremal coefficients}
\usage{
plotEC(matT, matE, flowConnect, ...)
}
\arguments{
\item{matT}{Matrix of parametric estimates of biavariate extremal coefficients}

\item{matE}{Matrix of non-parametric estimates of bivariate extremal coefficients}

\item{flowConnect}{Matrix of ones and zeros: 1 at row i and column j means that node i and node j are
flow connected}

\item{...}{additional arguments}
}
\description{
It generates a scatterplot of bivariate parametric versus non-parametric extremal coefficients.
It marks which ones extremal coefficients are between flow connected nodes. Used for application on rivers.
}
\note{
See Vignettes "Application Danube" and "Application Seine" for examples on post-estimation analysis.
}
