% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_pot.R
\name{pot.Network}
\alias{pot.Network}
\alias{pot.HRMtree}
\title{Peaks over threshold graph and data}
\usage{
\method{pot}{Network}(obj, obj2, k_ratio, ...)

\method{pot}{HRMtree}(obj, obj2, ...)
}
\arguments{
\item{obj}{Object of class \code{GTree} or \code{CovSelectTree} or \code{HRMtree}}

\item{obj2}{Object of class \code{RootDepSet}. If multiple subsets are passed only the first is considered.}

\item{k_ratio}{The number of the upper order statistics as a percentage of the sample size}
}
\value{
An object of the same class attribute as \code{obj} with two slots:

\code{g} is the induced subgraph on the vertices in \code{obj2}

\code{data} is the Peaks over threshold for the variables in the subset of \code{obj2}
}
\description{
Peaks over threshold graph and data
}
