% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_dataSets.R
\docType{data}
\name{Seine}
\alias{Seine}
\title{Seine Dataset}
\description{
Dataset of events on Seine, France. For detailed description of the data please read in the Appendix
of the paper below.
}
\references{
Asenova, S., Mazo, G. & Segers, J. Inference on extremal dependence in the domain of attraction of a
structured Huesler-Reiss distribution motivated by a Markov tree with latent variables.
Extremes 24, 461-500 (2021). https://doi.org/10.1007/s10687-021-00407-5
}
\author{
Stefka Asenova, Gildas Mazo, Johan Segers
}
\keyword{data}
