% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_setters.R
\name{setGraph.default}
\alias{setGraph.default}
\title{Default method for \code{setGraph}}
\usage{
\method{setGraph}{default}(obj, ...)
}
\value{
Returns warning if an unrecognised object is passed as an argument

It returns the argument that was called on
}
\description{
Default method for \code{setGraph}
}
