% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_setters.R
\name{setGraph.Network}
\alias{setGraph.Network}
\title{Setter for class \code{Network}}
\usage{
\method{setGraph}{Network}(obj, x, ...)
}
\arguments{
\item{obj}{Object of class \code{Network}}
}
\value{
The object
}
\description{
Setter for class \code{Network}
}
