% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_setters.R
\name{setRootDepSet}
\alias{setRootDepSet}
\title{Setter for \code{RootDepSet}}
\usage{
setRootDepSet(obj, subset, root, ...)
}
\arguments{
\item{obj}{Object of class \code{RootDepSet}}

\item{subset}{One dimensional array or list with subsets on the set of vertices, referred by their names.}

\item{root}{One dimensional array or a scalar of the root(s), corresponding to the subset(s) above
referred by name.}
}
\description{
It sets the values of the two slots in the object.
}
\examples{
rds<- RootDepSet()
rds_values<- list(a=c("a", "b"), b=c("b", "c", "d"), c=c("c", "d"), d=c("c", "d", "e"), e=c("d", "e"))
rds_roots<- c("a", "b","c","d","e")
rds<- setRootDepSet(rds,rds_values, rds_roots )
}
