% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_mySigma.R
\name{sigma.CovSelectTree}
\alias{sigma.CovSelectTree}
\title{Covariance matrix}
\usage{
\method{sigma}{CovSelectTree}(obj, s, set, ...)
}
\arguments{
\item{S_hat}{Empirical covariance matrix, not to be passed by user}
}
\description{
Covariance matrix
}
