% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_mySigma.R
\name{sigma}
\alias{sigma}
\title{Covariance matrix}
\usage{
sigma(obj, ...)
}
\arguments{
\item{obj}{Object of class \code{GTree} or \code{CovSelectTree} or \code{HRMtree}}

\item{...}{additional arguments}
}
\description{
Covariance matrix
}
\note{
It is not to be used independently
}
