% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_Simulation.R
\name{Simulation}
\alias{Simulation}
\title{Creates an object of class \code{Simulation}}
\usage{
Simulation(T, k, params, ends_to_params, estimator_names = NULL)
}
\arguments{
\item{T}{a matrix with four dimensions where the first dimension is the number of parameters,
the second the number of k used in the simulation, the third the number of estimators and the
fourth the number of repetitions in the simulation.}

\item{k}{a vector of k values used in the simulation}

\item{params}{a vector of numeric values of the true parameters}

\item{ends_to_params}{a matrix of dimensions #parameters*2 containing the incident vertices to
the edges.
it is assumed that the order of the parameters in T, trueParameters and endsToParameters}

\item{estimator_names}{a vector of strings indicating the names of the estimators used in
the simulation}
}
\description{
It is used to generate plots by using method \code{plot} on an object of that class.
}
