% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_auxiliaryFunctions.R
\name{spanned_subgraph}
\alias{spanned_subgraph}
\title{Creates a connected induced subgraph on a set of vertices}
\usage{
spanned_subgraph(g, u, W)
}
\arguments{
\item{g}{The original graph, an \code{igraph} object}

\item{u}{The root from which the paths to the other nodes will be computed}

\item{W}{The set of nodes to which the paths from the root will be computed}
}
\value{
A subgraph that is induced from all nodes on the paths between the root and the given set
of vertices. If the original graph is connected, then the spanned subgraph will be connected too.
}
\description{
On the basis of a root and a set of vertices it creates the tree spanned by the paths between
the root and the vertices.
}
