% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_subset.R
\name{subset.FlowConnect}
\alias{subset.FlowConnect}
\title{Subset from flow connection}
\usage{
\method{subset}{FlowConnect}(
  obj,
  from,
  g,
  vertices = get.vertex.attribute(g, "name", V(g)),
  matrixForm = FALSE,
  ...
)
}
\arguments{
\item{obj}{Object of class \code{FlowConnect}}

\item{from}{The object from which the subset should be extracted. For now two input types
are supported: \code{FlowConnectionMatrix} and \code{FlowConnectionGraph}.
In case of a matrix: entry ij is equal to 1 if node i is flow connected to node j.
In case of a graph: the graph should be directed according to the flow direction.}

\item{g}{\code{igraph} object, it may be directed (according to the flow) or undirected graph}

\item{vertices}{The vertices for which the set(s) of flow connected nodes are computed. The default
is all vertices in the graph}

\item{matrixForm}{Display results as matrix; default form}
}
\value{
Object of the same class as \code{obj} but with non empty slots

value A matrix or a list. If matrix the number of columns should be the same as the number of vertices
in the graph. If list the list should contain for each of the roots an array of the nodes with
which it is (a given root) flow connected.

root The vertices for which the flow connection sets are obtained
}
\description{
It derives the sbset(s) based on flow connection information.
}
\examples{
rdsobj<- FlowConnect()
seg<- graph(c(1,2, 2,3, 2,4, 4,5, 5,6, 5,7), directed = FALSE)
name_stat<- c("paris", "2", "meaux", "melun", "5", "nemours", "sens")
seg<- set.vertex.attribute(seg, "name", V(seg), name_stat)
seg_dir<- graph(c("2","paris", "meaux", "2", "melun", "2","5", "melun","nemours", "5", "sens", "5"), directed = TRUE)
fcg<- FlowConnectionGraph(seg_dir)
rdsobj<- subset(rdsobj, from=fcg, g=seg)
}
