% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_subset.R
\name{subset.Neighborhood}
\alias{subset.Neighborhood}
\title{Subset from neighborhood}
\usage{
\method{subset}{Neighborhood}(
  obj,
  eta,
  g,
  U_bar = NULL,
  vertices = base::setdiff(get.vertex.attribute(g, "name", V(g)), U_bar),
  matrixForm = FALSE,
  ...
)
}
\arguments{
\item{obj}{Object of class \code{Neighborhood}}

\item{eta}{Order of the neighborhood}

\item{g}{The graph, an \code{igraph} object}

\item{U_bar}{The set of nodes for which data are missing.}

\item{vertices}{Vertex(vertices) for which the neighbouring set(s) are derived. Default is all vertices.}
}
\value{
An object of class \code{Neighborhood} containing two slots: slot 'value' can be a matrix or list. If it
is a matrix, the number of columns should be the same as the number of vertices
in the graph. The number of columns should be the number of the parameter 'vertices'.
If it is a list, the list should contain for each of the nodes in the parameter 'vertices'
the neighborhood of order eta. The slot 'root' contains the vertices
for which the neighborhood sets are obtained.
}
\description{
Derives the subset based on neighborhood of a certain order
}
\examples{
seg<- graph(c(1,2, 2,3, 2,4, 4,5, 5,6, 5,7), directed = FALSE)
name_stat<- c("paris", "2", "meaux", "melun", "5", "nemours", "sens")
seg<- set.vertex.attribute(seg, "name", V(seg), name_stat)
rdsobj<- Neighborhood()
rdsobj<- subset(rdsobj, 2, seg)
}
