% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_subset.R
\name{subset}
\alias{subset}
\title{Generic for \code{subset}}
\usage{
subset(obj, ...)
}
\arguments{
\item{obj}{object of appropriate class}

\item{...}{additional arguments}
}
\description{
Generic for \code{subset}
}
\section{See \code{subset.FlowConnect} }{

For description and examples of the method applied on objects of class \code{FlowConnect}.
}

\section{See \code{subset.Neighborhood} }{

For description and examples of the method applied on objectobjects of class \code{Neighborhood}.
}

