% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_auxiliaryFunctions.R
\name{tilde_Rj}
\alias{tilde_Rj}
\title{Mean and covariance matrix}
\usage{
tilde_Rj(Ju, j, mu, sig)
}
\description{
Computes the mean and the covariance matrix for the vector \eqn{R_j}. It is designed to
be used by the method \code{stdf.EKS()}
}
