% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_validate.R
\name{validate}
\alias{validate}
\title{Generic function for method \code{validate}}
\usage{
validate(obj, ...)
}
\arguments{
\item{obj}{The object that needs to be validated}

\item{...}{Additional arguments}
}
\description{
Creates a generic function that contains a validator for an object
}
