% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_Argument.R
\name{Argument}
\alias{Argument}
\alias{ArgumentHvec}
\alias{ArgumentD}
\alias{ArgumentCC}
\title{Object with zero entries and of fixed dimensions}
\usage{
Argument(dims)

ArgumentHvec(dims)

ArgumentD(dims)

ArgumentCC(dims)
}
\description{
Definition of the class. \code{Argument} is an abstract class. The objects of this class
are used usually in the final estimation steps.

Class definition
It should be used when two dimensional matrices are to be halfvectorized
}
\details{
Creates a numeric object with given dimensions and zero entries
}
