% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_estimate.R
\name{estimate}
\alias{estimate}
\alias{estimate.HRMtree}
\alias{estimate.HRMBG}
\alias{estimate.MLE2}
\alias{estimate.EKS}
\alias{estimate.MMEave}
\alias{estimate.MLEave}
\alias{estimate.EngHitz}
\title{Provides estimation of a parametric model}
\usage{
estimate(obj, ...)

\method{estimate}{HRMtree}(obj, Data, subsets, k_ratio, xx = NULL, ...)

\method{estimate}{HRMBG}(obj, Data, subsets, k_ratio, xx = NULL, ...)

\method{estimate}{MLE2}(obj, Data, subsets, k_ratio, ...)

\method{estimate}{EKS}(obj, Data, coord, k_ratio, ...)

\method{estimate}{MMEave}(obj, Data, rdsobj, k_ratio, ss_bar = NULL, ...)

\method{estimate}{MLEave}(obj, Data, rdsobj, k_ratio, lmbd = NULL, ...)

\method{estimate}{EngHitz}(obj, data, subsets, k_ratio, ...)
}
\arguments{
\item{obj}{Object of appropriate class}

\item{Data}{Dataset with the variables placed on the columns, object of classes \code{matrix} or \code{data.frame}.
Each variable should have the same names as the node to which it belongs. For instance
if the name of the variable in the dataset is 'X2' the name of the node to which it belongs should be
'X2'.}

\item{subsets}{Object of class \code{RootDepSet}}

\item{k_ratio}{The fraction of the upper order statistics of the sample size}

\item{xx}{A named vector of coordinates, used to evaluate the stable tail dependence function.
The vector should be of length at least the number of vertices for which there is data available.
It should be named with the names of the vertices. The values given can be arbitrary, but values of
1 correspond to extremal coefficints.}

\item{coord}{A matrix of evaluation points with column names according to the
names of the edges. For creation of coordinates based on tuples, triples, and adjacent nodes see classes
\code{Coordinates}, \code{Tuples}, \code{Triples}, \code{Adjacent} as well as method \code{evalPoints}}
}
\value{
The estimates of \eqn{\theta}. For objects \code{MME} and \code{MLE} it returns the squared
values of the parameters.
}
\description{
Estimates the parameters of a statistical graphical model
}
