% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_flowConnectionInput.R
\name{FlowConnectionMatrix}
\alias{FlowConnectionMatrix}
\title{Object of class \code{FlowConnectionMatrix}}
\usage{
FlowConnectionMatrix(x, g)
}
\arguments{
\item{x}{should be a matrix of ones and zeros, 1 representing flow connected ness between node i and node j.
Note that x can not be symmetric.}

\item{g}{is an igraph object representing a directed graph corresponding to the flow connectedness.}
}
\description{
It is used as an input to create subsets based on the criterion of flow connectedness.
Within a subset all nodes will be flow connected.
}
