% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_getters.R
\name{getNoDataNodes}
\alias{getNoDataNodes}
\alias{getNoDataNodes.default}
\alias{getNoDataNodes.Network}
\title{It retrives the nodes without data}
\usage{
getNoDataNodes(object, ...)

\method{getNoDataNodes}{default}(object, ...)

\method{getNoDataNodes}{Network}(object, ...)
}
\arguments{
\item{obj}{Object of class 'Network'}
}
\description{
It retrieves the value of the slot 'noDataNodes', the set of nodes for which there are no data available
 for an object 'Network'.
}
