% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_getters.R
\name{getNodesWithData}
\alias{getNodesWithData}
\alias{getNodesWithData.default}
\alias{getNodesWithData.Network}
\title{It retrieves the nodes for which data are missing}
\usage{
getNodesWithData(obj, ...)

\method{getNodesWithData}{default}(obj, ...)

\method{getNodesWithData}{Network}(obj, ...)
}
\arguments{
\item{obj}{Object of class 'Network' or its subclasses \code{Tree, GTree, CovSelectTree, BlockGraph}}

\item{...}{additional arguments}
}
\description{
It retrieves the value of the slot \code{$nodesWithData}, the set of nodes for which there are data available
 for an object of class \code{Network}.
}
