% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_getters.R
\name{getParams}
\alias{getParams}
\alias{getParams.default}
\alias{getParams.HRMnetwork}
\title{It retrieves the parameters associated to an HRMnetwork}
\usage{
getParams(obj, ...)

\method{getParams}{default}(obj, ...)

\method{getParams}{HRMnetwork}(obj, ...)
}
\arguments{
\item{obj}{An object of class 'HRMnetwork'}

\item{...}{additional arguments}
}
\value{
The content of the slot \code{$depParams} of \code{obj}.
}
\description{
It retrieves the value of the slot \code{$depParams} of an object of class \code{HRMnetwork}.
}
