% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{HRMnetwork}
\alias{HRMnetwork}
\title{Object of class \code{HRMnetwork}}
\usage{
HRMnetwork(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{HRMnetwork} with two slots: \code{$graph} containing the graph and
\code{$depParams} the edge weights corresponding to the dependence parameters.
The dependence parameters are initialized at zero.
}
\description{
Object of class \code{HRMnetwork}
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
hrmnet<- HRMnetwork(g)
}
