% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{HRMtree}
\alias{HRMtree}
\title{\code{HRMtree} class}
\usage{
HRMtree(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{HRMtree} with two slots: one containing the graph and the other the edge weights
corresponding to the dependence parameters.
}
\description{
\code{HRMtree} class
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
hrmtree<- HRMtree(g)
}
