% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_HRMnetwork.R
\name{MMEave}
\alias{MMEave}
\title{Object of class \code{MMEave}}
\usage{
MMEave(g)
}
\arguments{
\item{g}{must be an \code{igraph} object}
}
\value{
Object of class \code{MMEave} with two slots: \code{$graph} containing the graph and \code{$depParams} containing the edge weights
corresponding to the dependence parameters.
}
\description{
It creates an object with two slots: \code{$graph} containing the graph and \code{$depParams} containing
the edge weights which are initialized with zero values.
It is intended to be used for estimation of edge weights of models on trees.
Consult Vignettes "Code - Note 6" and "Estimation - Note 6".
}
\examples{
g<- graph(c("a","b", "b","c", "b", "d"), directed=FALSE)
mmeave<- MMEave(g)
}
