% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_RootDepSet.R
\name{Neighborhood}
\alias{Neighborhood}
\title{An empty object of class \code{Neighborhood}}
\usage{
Neighborhood()
}
\arguments{
\item{value}{Matrix or list containing subsets of the vertex set.
A subset represents a neighborhood of a given node.}

\item{root}{A list of vertices.}
}
\description{
An empty object of class \code{Neighborhood}
}
\examples{
seg<- graph(c(1,2, 2,3, 2,4, 4,5, 5,6, 5,7), directed = FALSE)
name_stat<- c("paris", "2", "meaux", "melun", "5", "nemours", "sens")
seg<- set.vertex.attribute(seg, "name", V(seg), name_stat)
rdsobj<- Neighborhood()
rdsobj<- subset(rdsobj, 2, seg)
}
