% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_Network.R
\name{Network}
\alias{Network}
\title{\code{Network} class}
\usage{
Network(x, data)
}
\arguments{
\item{x}{An object of class \code{igraph}}

\item{data}{Data set associated to the graph}
}
\value{
An object of class \code{Network}
}
\description{
Class definition
\code{Network} is a superclass
}
