% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_RootDepSet.R
\name{RootDepSet}
\alias{RootDepSet}
\title{An empty object of class \code{RootDepSet}}
\usage{
RootDepSet()
}
\arguments{
\item{value}{Matrix or list containing subsets of the vertex set}

\item{root}{A list of vertices.}
}
\description{
An empty object of class \code{RootDepSet}
}
\examples{
rdsobj<- RootDepSet()
rds_values<- list(a=c("a", "b"), b=c("b", "c", "d"), c=c("c", "d"), d=c("c", "d", "e"), e=c("d", "e"))
rds_roots<- c("a", "b","c","d","e")
rdsobj<- setRootDepSet(rdsobj, rds_values, rds_roots)
}
