% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_setters.R
\name{setParams}
\alias{setParams}
\title{Sets the parameters attached to the edges of the tree}
\usage{
setParams(obj, value, ...)
}
\arguments{
\item{value}{A named vector with names corresponding to the names of the edges}
}
\description{
Sets the parameters attached to the edges of the tree
}
