% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_mySigma.R
\name{sigma}
\alias{sigma}
\title{Covariance matrix}
\usage{
sigma(obj, ...)
}
\arguments{
\item{obj}{Object of class \code{GTree} or \code{HRMtree}}

\item{...}{additional arguments}
}
\value{
A matrix of non-parametric estimates or a matrix of zeros and ones, say A, such that A*theta is
the halfvectorized Sigma,
 where Sigma is the parametric covariance matrix.
 If the first argument is of class \code{GTree} the method returns the empirical covariance matrix
 of the data in the GTree object.
If the first argument is of class \code{HRMtree} the method returns
a matrix of ones and zeros which corresponds to \eqn{hvec(\Sigma(\theta))=A\theta}
}
\description{
It computes either a non-parametric estimate of the covariance matrix or a parametric form of the
covariance matrix.
}
\note{
It is not intendend for independent use.
}
