% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_mySigma.R
\name{sigma}
\alias{sigma}
\alias{sigma.Network}
\alias{sigma.HRMtree}
\title{Covariance matrix}
\usage{
sigma(obj, ...)

\method{sigma}{Network}(obj, ...)

\method{sigma}{HRMtree}(obj, obj2, U_bar = NULL, ...)
}
\arguments{
\item{obj}{Object of class \code{GTree} or \code{CovSelectTree} or \code{HRMtree}}

\item{obj2}{Object of class \code{RootDepSet}}

\item{U_bar}{Vector of the missing vertices in a graph, the default is NULL}
}
\value{
see Details.
}
\description{
It computes the empirical or the parametric covariance matrix depending on the class of
its first argument. See Details.
}
\details{
If the first argument is of class \code{GTree} the method returns the empirical covariance matrix
of the data in the GTree object.

If the first argument is of class \code{CovSelectTree} the method returns the covariance matrix
of the corresponding Gaussian graphical model associated to the object.

If the first argument is of class \code{HRMtree} the method returns
a matrix of ones and zeros which corresponds to \eqn{hvec(\Sigma(\theta))=A\theta}
}
