% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_stdf.R
\name{stdf}
\alias{stdf}
\alias{stdf.Network}
\alias{stdf.EKS}
\title{Stable tail dependence function}
\usage{
stdf(obj, ...)

\method{stdf}{Network}(obj, evalPoints, k_ratio)

\method{stdf}{EKS}(obj, x, Ubar = NULL)
}
\arguments{
\item{obj}{An object of the corresponding class}

\item{evalPoints}{A matrix of evaluation points with named columns, dimensions qx|U| where
U is the set of nodes with available data}

\item{k_ratio}{The share of the largest order statistics}

\item{x}{A named vector of evaluation points}

\item{Ubar}{The set of nodes for which data are missing}
}
\description{
Computes the stable tail dependence function for a given matrix of evaluation points
}
