% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_Network.R
\name{Tree}
\alias{Tree}
\title{\code{Tree} class}
\usage{
Tree(x, data)
}
\arguments{
\item{x}{An object of class \code{igraph}}

\item{data}{Data set associated to the graph}
}
\value{
An object of class \code{Tree}
}
\description{
Class definition
}
